Further DRY out test code
Separate the test code that is shared between both scenarios from the
code that is unique. The shared test code is symlinked from the default
scenario into the with-texmf-options scenario. This will help ensure
that changes to shared test code are consistent between the two
scenarios.
